@@ -16,15 +16,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
 
     defaultConfig {
         applicationId "com.ashomok.lullabies"
         minSdkVersion 17
-        targetSdkVersion 28
-        versionCode 18
-        versionName "3.3"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 29
+        versionCode 19
+        versionName "3.4"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
     }
     compileOptions {
@@ -51,16 +51,16 @@ repositories {
 dependencies {
     compileOnly 'com.google.android.wearable:wearable:2.4.0'
 
-    implementation 'com.google.android.gms:play-services-cast-framework:16.2.0'
+    implementation 'com.google.android.gms:play-services-cast-framework:17.0.0'
     implementation 'com.google.android.support:wearable:2.4.0'
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:cardview-v7:28.0.0'
-    implementation 'com.android.support:mediarouter-v7:28.0.0'
-    implementation 'com.android.support:leanback-v17:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
-    implementation 'com.android.support:percent:28.0.0'
-    implementation 'com.google.firebase:firebase-ads:17.2.1'
-    implementation 'com.android.support:multidex:1.0.3'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.mediarouter:mediarouter:1.0.0'
+    implementation 'androidx.leanback:leanback:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.percentlayout:percentlayout:1.0.0'
+    implementation 'com.google.firebase:firebase-ads:18.0.0'
+    implementation 'androidx.multidex:multidex:2.0.1'
     implementation 'com.android.billingclient:billing:1.0'
 
     // Dagger docs https://dagger.dev/
@@ -72,7 +72,7 @@ dependencies {
     implementation 'com.annimon:stream:1.1.8'
 
     implementation 'com.google.android.exoplayer:exoplayer:2.7.3'
-    implementation 'com.google.firebase:firebase-core:16.0.9'
+    implementation 'com.google.firebase:firebase-core:17.0.0'
 
     //rx
     implementation group: 'io.reactivex.rxjava2', name: 'rxjava', version: '2.1.3'
@@ -88,17 +88,17 @@ dependencies {
     androidTestImplementation 'org.mockito:mockito-core:2.28.2'
     androidTestImplementation 'org.mockito:mockito-android:2.28.2'
     // Espresso UI Testing
-    androidTestImplementation "com.android.support.test.espresso:espresso-core:3.0.2"
-    androidTestImplementation "com.android.support.test.espresso:espresso-contrib:3.0.2"
-    androidTestImplementation "com.android.support.test.espresso:espresso-intents:3.0.2"
-    androidTestImplementation "com.android.support.test.espresso.idling:idling-concurrent:3.0.2"
-    implementation "com.android.support.test.espresso:espresso-idling-resource:3.0.2"
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-contrib:3.2.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-intents:3.2.0'
+    androidTestImplementation 'androidx.test.espresso.idling:idling-concurrent:3.2.0'
+    implementation 'androidx.test.espresso:espresso-idling-resource:3.2.0'
 
 
     androidTestImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support:support-annotations:28.0.0'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test:rules:1.0.2'
+    androidTestImplementation 'androidx.annotation:annotation:1.1.0'
+    androidTestImplementation 'androidx.test:runner:1.2.0'
+    androidTestImplementation 'androidx.test:rules:1.2.0'
 }
 
 apply plugin: 'com.google.gms.google-services'
