@@ -1,7 +1,7 @@
 package com.ashomok.lullabies.ad;
 
 import android.content.Context;
-import android.support.annotation.StringRes;
+import androidx.annotation.StringRes;
 import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
