@@ -1,9 +1,9 @@
 package com.ashomok.lullabies.billing;
 
 import android.app.Activity;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.annotation.StringRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.StringRes;
 import android.util.Log;
 
 import com.android.billingclient.api.BillingClient;
