@@ -1,8 +1,8 @@
 package com.ashomok.lullabies.utils;
 
-import android.support.annotation.StringRes;
-import android.support.design.widget.Snackbar;
-import android.support.v4.content.ContextCompat;
+import androidx.annotation.StringRes;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.core.content.ContextCompat;
 import android.view.View;
 
 import com.ashomok.lullabies.R;
